# src/app/api/crm/members/export/route.ts
import { NextRequest, NextResponse } from 'next/server';

import type { ExportMembersRequest } from '@/types/member.type';

export async function POST(request: NextRequest) {
  try {
    const body: ExportMembersRequest = await request.json();

    // Validate export count (max 10,000)
    if (body.target === 'filtered' && body.memberIds && body.memberIds.length > 10000) {
      return NextResponse.json({ error: 'Export limit is 10,000 members' }, { status: 400 });
    }

    // Validate fields
    if (!body.fields || body.fields.length === 0) {
      return NextResponse.json({ error: 'At least one field must be selected' }, { status: 400 });
    }

    // Mock export response
    return NextResponse.json({
      success: true,
      exportId: `export-${Date.now()}`,
      format: body.format,
      status: 'processing',
    });
  } catch (error) {
    return NextResponse.json({ error: 'Failed to export members' }, { status: 500 });
  }
}