# src/lib/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fitness CRM API",
    "version": "1.0.0",
    "description": "API documentation for Fitness CRM system"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/crm/members/{id}/points": {
      "get": {
        "tags": ["Members"],
        "summary": "Get member points",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": ["Crm"],
        "summary": "POST /crm/members/{id}/points",
        "description": "Auto-generated from src/app/api/crm/members/[id]/points/route.ts",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id parameter"
          }
        ]
      }
    },
...
    "/crm/members/export": {
      "post": {
        "tags": ["Members"],
        "summary": "Export members",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportMembersRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/crm/members/{id}/memos/{memoId}": {
      "put": {
        "tags": ["Crm"],
        "summary": "PUT /crm/members/{id}/memos/{memoId}",
        "description": "Auto-generated from src/app/api/crm/members/[id]/memos/[memoId]/route.ts",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id parameter"
          },
          {
            "name": "memoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "memoId parameter"
          }
        ]
      },
      "delete": {
        "tags": ["Crm"],
        "summary": "DELETE /crm/members/{id}/memos/{memoId}",
        "description": "Auto-generated from src/app/api/crm/members/[id]/memos/[memoId]/route.ts",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id parameter"
          },
          {
            "name": "memoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "memoId parameter"
          }
        ]
      }
    },
...
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "POST /auth/register",
        "description": "Auto-generated from src/app/api/auth/register/route.ts",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ExportMembersRequest": {
        "type": "object",
        "required": ["format", "target", "fields"],
        "properties": {
          "format": {
            "type": "string",
            "enum": ["csv", "excel"]
          },
          "target": {
            "type": "string",
            "enum": ["selected", "filtered"]
          },
          "memberIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}